JAMES-2341 Move SpamAssassinListenerModule to its own project
