+++ server/container/guice/mailbox-plugin-spamassassin/src/main/java/org/apache/james/modules/spamassassin/SpamAssassinListenerModule.java
@@ -17,27 +17,33 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.modules.mailbox;
+package org.apache.james.modules.spamassassin;
 
 import java.io.FileNotFoundException;
+import java.util.List;
 
 import javax.inject.Singleton;
 
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.PropertiesConfiguration;
+import org.apache.james.lifecycle.api.Configurable;
 import org.apache.james.mailbox.MailboxSession;
 import org.apache.james.mailbox.exception.MailboxException;
 import org.apache.james.mailbox.spamassassin.SpamAssassin;
 import org.apache.james.mailbox.spamassassin.SpamAssassinConfiguration;
 import org.apache.james.mailbox.spamassassin.SpamAssassinListener;
 import org.apache.james.mailbox.store.StoreMailboxManager;
-import org.apache.james.mailbox.store.event.SpamEventListener;
+import org.apache.james.utils.ConfigurationPerformer;
 import org.apache.james.utils.PropertiesProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.base.Throwables;
+import com.google.common.collect.ImmutableList;
 import com.google.inject.AbstractModule;
+import com.google.inject.Inject;
 import com.google.inject.Provides;
+import com.google.inject.multibindings.Multibinder;
 
 public class SpamAssassinListenerModule extends AbstractModule {
     private static final Logger LOGGER = LoggerFactory.getLogger(SpamAssassinListenerModule.class);
@@ -46,15 +52,36 @@ public class SpamAssassinListenerModule extends AbstractModule {
 
     @Override
     protected void configure() {
+        Multibinder.newSetBinder(binder(), ConfigurationPerformer.class).addBinding().to(SpamAssassinListenerConfigurationPerformer.class);
     }
-
-    @Provides
+    
     @Singleton
-    public SpamEventListener provideSpamEventListener(SpamAssassinConfiguration spamAssassinConfiguration, StoreMailboxManager storeMailboxManager) throws MailboxException {
-        SpamAssassinListener spamAssassinListener = new SpamAssassinListener(new SpamAssassin(spamAssassinConfiguration));
-        MailboxSession session = null;
-        storeMailboxManager.addGlobalListener(spamAssassinListener, session);
-        return spamAssassinListener;
+    public static class SpamAssassinListenerConfigurationPerformer implements ConfigurationPerformer {
+
+        private final SpamAssassinConfiguration spamAssassinConfiguration;
+        private final StoreMailboxManager storeMailboxManager;
+
+        @Inject
+        public SpamAssassinListenerConfigurationPerformer(SpamAssassinConfiguration spamAssassinConfiguration, StoreMailboxManager storeMailboxManager) {
+            this.spamAssassinConfiguration = spamAssassinConfiguration;
+            this.storeMailboxManager = storeMailboxManager;
+        }
+
+        @Override
+        public void initModule() {
+            try {
+                SpamAssassinListener spamAssassinListener = new SpamAssassinListener(new SpamAssassin(spamAssassinConfiguration));
+                MailboxSession session = null;
+                storeMailboxManager.addGlobalListener(spamAssassinListener, session);
+            } catch (MailboxException e) {
+                Throwables.propagate(e);
+            }
+        }
+
+        @Override
+        public List<Class<? extends Configurable>> forClasses() {
+            return ImmutableList.of();
+        }
     }
 
     @Provides
